refactor (home page): Implemented Atomic Component Architecture and SEO Compliance
This commit introduces a significant restructuring of the Frontend directory to improve code organization and maintainability. Key changes include:

- **Component Restructuring:**
    - `footer.jsx` moved to `Frontend/src/components/Footer/footer.jsx`
    - `navbar.jsx` moved to `Frontend/src/components/Header/Header.jsx`
- **Page Restructuring:**
    - The `home.jsx` page has been refactored into a dedicated `Home` directory (`Frontend/src/pages/Home/`) to house individual sections of the homepage.
    - New components created for each section of the homepage:
        - `HeroSection.jsx`
        - `AboutSection.jsx`
        - `EventsSection.jsx`
        - `MediaSection.jsx`
        - `NewsletterSection.jsx`
        - `NotesSection.jsx`
    - A new `Home.jsx` file now serves as the main entry point for the homepage, integrating all these sections.
- **Asset Renaming:**
    - `finanza-club-logo.png` renamed to `finanza-club-square-logo.png` for clarity.
    - `FooterImg.png` renamed to `finanza-footer.png` for consistency.
- **Dependency Updates:**
    - Using _react-icons_ instead of _lucid icons_
    - `package.json` and `package-lock.json` have been updated to reflect any new or modified dependencies.
- **Minor Changes:**
    - `.DS_Store` and `index.html` have been modified.
    - `favicon.png` has been deleted.

This refactoring aims to make the codebase more modular, easier to navigate, and more scalable for future development.
