+++ Frontend/index.html
@@ -2,9 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./public/finanza-club-logo.png" />
+    <meta name="description" content="Finanza Club – The Finance & Investment Club at Rishihood University">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Finanza</title>
+    <link rel="icon" type="image/svg+xml" href="./finanza-club-square-logo.png" />
+    <title>Finanza Club – The Finance & Investment Club at Rishihood University</title>
   </head>
   <body>
     <div id="root"></div>
