+++ Frontend/package.json
@@ -14,6 +14,7 @@
     "lucide-react": "^0.553.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^7.9.5"
   },
   "devDependencies": {
