+++ Frontend/src/components/Header/Header.jsx
@@ -1,12 +1,19 @@
 import { Link } from "react-router-dom";
+import { FaSearch } from "react-icons/fa";
 
-export default function Navbar() {
+export default function Header() {
   return (
-    <nav className="bg-white shadow-sm border-b border-gray-200">
-      <div className="container mx-auto px-8 py-3">
+    <header className="bg-white shadow-sm border-b border-gray-200">
+      <nav className="container mx-auto px-8 py-3">
         <div className="flex justify-between items-center">
           {/* Logo with border box */}
-          <div >
+          <div className="flex items-center gap-1">
+            <img
+              className="w-10 h-full" 
+              src="./finanza-club-square-logo.png" 
+              alt="Finanza Club" 
+              draggable="false"
+            />
             <h1 className="font-bold text-3xl tracking-tight" style={{ color: '#8B1538' }}>
               Finanza Club 
             </h1>
@@ -16,31 +23,31 @@ export default function Navbar() {
           <div className="flex items-center gap-10 font-medium text-base">
             <Link
               to="/events"
-              className="text-gray-800 hover:text-gray-600 transition-colors duration-200"
+              className="text-gray-800 hover:text-[#C0003D] transition-colors duration-200"
             >
               Events
             </Link>
             <Link
               to="/newsletters"
-              className="text-gray-800 hover:text-gray-600 transition-colors duration-200"
+              className="text-gray-800 hover:text-[#C0003D] transition-colors duration-200"
             >
               Newsletters
             </Link>
             <Link
               to="/"
-              className="text-gray-800 hover:text-gray-600 transition-colors duration-200"
+              className="text-gray-800 hover:text-[#C0003D] transition-colors duration-200"
             >
               Blogs
             </Link>
             <Link
               to="/members"
-              className="text-gray-800 hover:text-gray-600 transition-colors duration-200"
+              className="text-gray-800 hover:text-[#C0003D] transition-colors duration-200"
             >
               Members
             </Link>
             <Link
               to="/about"
-              className="text-gray-800 hover:text-gray-600 transition-colors duration-200"
+              className="text-gray-800 hover:text-[#C0003D] transition-colors duration-200"
             >
               About
             </Link>
@@ -53,9 +60,11 @@ export default function Navbar() {
             >
               Contact us
             </Link>
+            <FaSearch className="w-4 h-4 "/>
+            {/* <button>Dark Mode</button> */}
           </div>
         </div>
-      </div>
-    </nav>
+      </nav>
+    </header>
   );
 }
