+++ Frontend/src/main.jsx
@@ -12,5 +12,3 @@ createRoot(document.getElementById('root')).render(
     </BrowserRouter>
   </StrictMode>,
 )
-
-
