+++ Frontend/src/pages/Home/AboutSection.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+function AboutSection() {
+    {/* About Section */}
+    return (
+        <section className="max-w-7xl mx-auto px-6 md:px-12 mb-20">
+            <h2 className="text-3xl font-bold mb-2">About <span className="text-3xl font-bold text-[#C0003D] mb-6">Finanza</span></h2>
+            {/* <span className="text-4xl font-bold text-[#C0003D] mb-6">Finanza</span> */}
+            <div className="text-gray-700 leading-relaxed space-y-4">
+                <p>
+                    Lorem ipsum dolor sit amet consectetur. Nec sollicitudin feugiat sed vel sit. Rhoncus nec in sollicitudin feugiat sed turpis ut arcu non risus. At tortor placerat at dignissim aliquet. Et facilisi vivamus diam quam interdum placerat non lobortis velit scelerisque. Eget venenatis sed vel sit sed. Ipsum sollicitudin feugiat lobortis lacinia. Tempor morbi bibendum pharetra velit turpis scelerisque sed. Sed habitasse viverra purus.
+                </p>
+                <p>
+                    Ultrices facilisi quis sem integer sem orci. Morbi sit tortor. Ultrices feugiat imperdiet lectis. Eget rhoncus non gravit placerat.
+                </p>
+            </div>
+        </section>
+    );
+}
+
+export default AboutSection;
